pyproject.toml
[tool.poetry]
name = "sam-lora-poetry"
version = "0.1.0"
description = ""
authors = ["WangRongsheng <wrs6@88.com>", "Ivan Serebrennikov <w7a8n1y4a@gmail.com>"]
readme = "README.md"
packages = [{include = "sam_lora_poetry"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pendulum = "^2.1.2"
numpy = "^1.26.0"
matplotlib = "^3.8.0"
torchvision = ">0.15.2"
statistics = "^1.0.3.5"
tqdm = "^4.66.1"
monai = "^1.2.0"
pillow = "^10.0.1"
safetensors = "^0.3.3"
opencv-python = "<=4.8.0.74"
pyyaml = "^6.0.1"
gradio = "^3.44.3"
torch = "2.1.2"
dvc = {extras = ["s3"], version = "^3.33.3"}
python-dotenv = "^1.0.1"
checksumdir = "^1.2.0"
httpx = "^0.27.0"
mlflow = "^2.11.3"
black = "^24.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
